requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Blockchain & Web3
web3==6.15.1
eth-account==0.11.0
eth-utils==4.0.0

# Payment Integration
stripe==8.0.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# WebSocket
websockets==12.0

# Data Validation & Serialization
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pydantic[email]==2.5.3

# CLI & UI
streamlit==1.30.0
rich==13.7.0
plotly==5.18.0
pandas==2.1.4

# Async Support
asyncio==3.4.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Logging
loguru==0.7.2
